requirements.txt
tensorflow==2.11.0
tensorflow-privacy==0.7.0
numpy>=1.19.2
scipy>=1.5.2
matplotlib>=3.3.2
tensorflow-datasets>=4.0.1
PyYAML>=5.4.1

flwr>=1.0.0
